server/requirements.txt
# רשימת הספריות שהפרויקט צריך בשביל לרוץ
flask
flask-sqlalchemy
flask-cors
pillow
werkzeug
opencv-python